requirements.txt
paho-mqtt
psycopg2
python-dotenv
schedule